requirements.txt
django
django-extra-views
